gunslinger/requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
APScheduler==3.6.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
boto3==1.12.31
botocore==1.15.31
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
docutils==0.15.2
fake-useragent==0.1.11
idna==2.9
jmespath==0.9.5
lxml==4.5.1
multidict==4.7.5
parse==1.15.0
pluginbase==1.0.0
pyee==7.0.2
pyppeteer==0.2.2
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
slackclient==2.5.0
soupsieve==2.0.1
tqdm==4.46.0
tzlocal==2.0.0
urllib3==1.24.3
w3lib==1.22.0
websockets==8.1
yarl==1.4.2
